adjust callgraph colors and spacing
